feat(web): Redesigned Satisfaction section
